sequencial http request

+++ src/main/kotlin/com/wenbiming/dopc/data/Delivery.kt
@@ -9,5 +9,5 @@ data class DeliveryResponse(
 
 data class Delivery(
     val fee: Int,
-    val distance: Double
+    val distance: Int
 )
